# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/countries:
    get:
      tags:
      - countries-controller
      operationId: search
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CountryTableDto"
  /api/countries/{iso3}:
    get:
      tags:
      - countries-controller
      operationId: getCountryByIso3
      parameters:
      - name: iso3
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CountryDetailDto"
components:
  schemas:
    CountryTableDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        iso3:
          type: string
        numericCode:
          type: integer
          format: int32
        capitalName:
          type: string
        currency:
          type: string
        region:
          type: string
        subregion:
          type: string
        latitudeLongitude:
          type: string
        emoji:
          type: string
    CityDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        latitude:
          type: integer
          format: int32
        longitude:
          type: integer
          format: int32
    CountryDetailDto:
      type: object
      properties:
        country:
          $ref: "#/components/schemas/CountryDto"
        states:
          uniqueItems: true
          type: array
          items:
            $ref: "#/components/schemas/StateDto"
    CountryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        iso2:
          type: string
        iso3:
          type: string
        numericCode:
          type: integer
          format: int32
        phoneCode:
          type: string
        capitalName:
          type: string
        currency:
          type: string
        currencyName:
          type: string
        currencySymbol:
          type: string
        tld:
          type: string
        nativeName:
          type: string
        region:
          type: string
        regionId:
          type: integer
          format: int32
        subregion:
          type: string
        subregionId:
          type: integer
          format: int32
        nationality:
          type: string
        timezones:
          type: string
        latitude:
          type: integer
          format: int32
        longitude:
          type: integer
          format: int32
        emoji:
          type: string
        emojiU:
          type: string
    StateDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        latitude:
          type: integer
          format: int32
        longitude:
          type: integer
          format: int32
        cities:
          type: array
          items:
            $ref: "#/components/schemas/CityDto"